Isolate method_missing functions into its own runtime module
